refactor(persis): relocate filenotification, fileincident, fileserviceregistry under internal/persis/file/
Move the three remaining file-backed multi-record leaves into persis/file/
alongside dagrun, proc, and session:

  filenotification     → file/notification    (multi-subdir + encryption + channels + routes)
  fileincident         → file/incident        (4 subdirs: providers, policy/workspaces, policy/dags, state)
  fileserviceregistry  → file/serviceregistry (registry + finder + cleaner background goroutine)

These were originally enumerated for porting to store/<name>.go adapters,
but their shapes (multi-record-type, multi-subdir, background routines)
do not fit the single-record-per-id Collection abstraction cleanly.
Forcing them in would introduce per-store Collection options and multi-
collection adapters that obscure the simple stores' pattern. Following
the dagrun and proc precedent: keep them as backend-specific file
implementations under file/, behind the same wiring boundary so callers
see no difference.

Consumer imports keep their original fileXxx alias so no caller code
changes.

Pure move: no logic, no test, no on-disk-format changes.
