+++ internal/persis/file/serviceregistry/cleaner_test.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2026 Yota Hamada
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-package fileserviceregistry
+package serviceregistry
 
 import (
 	"context"
