+++ internal/service/frontend/api/v1/notification_channels_license_test.go
@@ -13,7 +13,7 @@ import (
 	dagucrypto "github.com/dagucloud/dagu/internal/cmn/crypto"
 	"github.com/dagucloud/dagu/internal/license"
 	notificationmodel "github.com/dagucloud/dagu/internal/notification"
-	"github.com/dagucloud/dagu/internal/persis/filenotification"
+	filenotification "github.com/dagucloud/dagu/internal/persis/file/notification"
 	"github.com/dagucloud/dagu/internal/service/eventstore"
 	"github.com/dagucloud/dagu/internal/service/frontend"
 	"github.com/dagucloud/dagu/internal/test"
